put popupPlugin in its own file
